Conditional add

+++ pupapply.sh
@@ -20,7 +20,9 @@ if [ ! -f "$MANIFEST_DIR/$MANIFEST" ] || [ ! "$(diff "$MANIFEST" "$MANIFEST_DIR/
 	cp "$MANIFEST" "$MANIFEST_DIR/$MANIFEST"
 fi
 
-cp "$0" "/root/"
+if [ ! -f "/root/pupapply.sh" ]; then
+	cp "$0" "/root/"
+fi
 
 if [ -d "$MODULE_DIR" ]; then
 	echo "Applying manifest."
